refactor: 사용 용도에 따른 components 디렉토리 refactor
